removed previous fetch codes
